completando parte del tp9

+++ Models/Productos.cs
@@ -3,5 +3,5 @@ public class Productos
 {
     public int idProducto { get; set; }
     public string descripcion { get; set; }
-    public int precio { get; set; }
+    public decimal precio { get; set; }
 }
